removed storybook sourcing as we won't use it
