+++ apps/swirl-docs/src/lib/docs/src/links.ts
@@ -1,6 +1,15 @@
 import { DocCategory, DocHeadline, DOCUMENTATION_SRC } from "./docs.model";
-import { createLinkListForDocument } from "./singleDoc";
 import { createDocCategory } from "./multipleDocs";
+import { generateSerializableDocumentation } from "./singleDoc";
+
+const HeadingMap = new Map<string, number>([
+  ["#", 1],
+  ["##", 2],
+  ["###", 3],
+  ["####", 4],
+  ["#####", 5],
+  ["######", 6],
+]);
 
 export function createLinkLists(
   category: string,
@@ -23,4 +32,28 @@ export function createLinkLists(
     documentLinkList,
     categoryLinkList,
   };
+}
+
+export function createLinkListForDocument(
+  category: string,
+  document: string
+): DocHeadline[] {
+  const source = generateSerializableDocumentation(category, document);
+  const headlines = source.split("\n").filter((line) => line.startsWith("#"));
+
+  return headlines.map((headline) => {
+    const headlineId = headline
+      .split(" ")
+      .slice(1, headline.length)
+      .join("-")
+      .toLowerCase();
+
+    const headlineLevel = HeadingMap.get(headline.split(" ")[0]);
+
+    return {
+      id: headlineId,
+      name: headline.split(" ")[1],
+      level: headlineLevel,
+    };
+  });
 }
